[AJ-1296] Move files out of src/data (#4160)
